Refactorización del método borrar
