+++ restapi/routes/operations/find_all.ts
@@ -2,7 +2,7 @@ import { Response } from "express";
 import * as service from "../../services/DB_manager";
 
 // Unifica todos los accesos para recuperar la collección entera de un determinado elemento
-export async function findAll(collectionName : string, res: Response) {
+export async function findAllDocuments(collectionName : string, res: Response) {
     try {
         var documents = await service.getCollection(collectionName); //Se obtienen los datos del servicio
