+++ restapi/routes/products_router.ts
@@ -4,14 +4,15 @@ import * as mongodb  from "mongodb";
 import * as service from "../services/DB_manager";
 import Product from "../models/product";
 import sanitizeHtml from "sanitize-html";
-import { findAll } from "./operations/find_all";
-import { update } from "./operations/update";
+import { findAllDocuments } from "./operations/find_all";
+import { updateDocument } from "./operations/update";
+import { deleteDocument } from "./operations/delete";
 
 var app = require("../server");
 
 // GET (todos los productos)
 app.get("/product/", async (_req: Request, res: Response) => {
-    findAll("Producto", res)
+    findAllDocuments("Producto", res)
 });
 
 //ByID
@@ -47,7 +48,7 @@ app.post("/product/", async (req: Request, res: Response) => {
 
 // PUT (update)
 app.put("/product/:id", async (req: Request, res: Response) => {
-    update(req, res,  //Llama a operations/update
+    updateDocument(req, res,  //Llama a operations/update
         async (id: string) =>{
             var updatedProduct: Product = req.body;
             var query = { _id: new mongodb.ObjectId(id) };
@@ -62,21 +63,7 @@ app.put("/product/:id", async (req: Request, res: Response) => {
 
 // DELETE
 app.delete("/product/:id", async (req: Request, res: Response) => {
-    var id = req?.params?.id;
-
-    try {
-        var query = { _id: new mongodb.ObjectId(id) };
-        var result = await service.removeProduct(query);
-
-        if (result && result.deletedCount) {
-            res.status(202).send(sanitizeHtml(`Successfully removed product with id ${id}`));
-        } else if (!result) {
-            res.status(400).send(sanitizeHtml(`Failed to remove product with id ${id}`));
-        } else if (!result.deletedCount) {
-            res.status(404).send(sanitizeHtml(`Product with id ${id} does not exist`));
-        }
-    } catch (error) {
-        console.error(error.message);
-        res.status(400).send(error.message);
-    }
+    deleteDocument(req, res, 
+        async (query:string) => { console.log(query)
+            return await service.removeProduct(query);})
 });
